fix:modify code difference

+++ framework/innerkits/include/pasteboard_web_controller.h
@@ -65,7 +65,7 @@ private:
     std::pair<std::shared_ptr<PasteDataRecord>, std::shared_ptr<std::string>> ExtractContent(
         const RecordList &records, OffsetMap &replaceUris);
     void UpdateHtmlRecord(std::shared_ptr<PasteDataRecord> &htmlRecord, std::shared_ptr<std::string> &htmlData);
-    int32_t GetNeedCheckUris(PasteData &pasteData, std::vector<std::string> &uris,
+    size_t GetNeedCheckUris(PasteData &pasteData, std::vector<std::string> &uris,
         std::vector<size_t> &indexs, bool ancoFlag);
     void SetUriPermission(std::shared_ptr<PasteDataRecord> &record, bool isRead, bool isWrite, bool isNeedPersistance);
 };
